Add static compile-time checking of printf functions in liboctave.
* util/lo-error.h: Use macro OCTAVE_FORMAT_PRINTF for printf functions.

* array/Array.h: Use format specifier "z" for type size_t.

* array/Array-util.cc, numeric/oct-norm.cc: Supply a format string of "%s" when
  printing a single string.

* array/idx-vector.cc, array/MatrixType.cc, array/Sparse.cc, numeric/DASPK.cc,
  numeric/DASRT.cc, numeric/DASSL.cc, numeric/LSODE.cc,
  util/lo-array-errwarn.cc, util/sparse-util.cc: Use preprocessor macro
  OCTAVE_IDX_TYPE_FORMAT as the format identifier for octave_idx_type.

* util/f77-fcn.c: Cast width argument to int.

* util/lo-utils.cc: Pass missing argument c0.

+++ liboctave/util/f77-fcn.c
@@ -53,7 +53,7 @@ F77_FUNC (xstopx, XSTOPX) (F77_CONST_CHAR_ARG_DEF (s_arg, len)
       slen = strlen (s);
     }
 
-  (*current_liboctave_error_handler) ("%.*s", slen, s);
+  (*current_liboctave_error_handler) ("%.*s", (int) slen, s);
 
   F77_NORETURN (0)
 }
